+++ src/task.js
@@ -1,6 +1,7 @@
-import createElement from './create-element.js';
-export default class Task {
+import Component from './component.js';
+export default class Task extends Component {
   constructor(data) {
+    super();
     this._title = data.title;
     this._dueDay = data.dueDay;
     this._tags = data.tags;
@@ -8,32 +9,22 @@ export default class Task {
     this._repeatingDays = data.repeatingDays;
     this._color = data.color;
 
-    this._element = null;
-    this._state = {
-      // isEdit: false
-    };
     this._onEdit = null;
+    this._onEditButtonClick = this._onEditButtonClick.bind(this);
 
   }
   _isRepeating() {
     return Object.values(this._repeatingDays).some((it) => it === true);
   }
   _onEditButtonClick() {
-    // this._state.isEdit = this._state.isEdit;
-    // this.update();
     return typeof this._onEdit === `function` && this._onEdit();
   }
-  get element() {
-    return this._element;
-  }
-
   set onEdit(fn) {
     this._onEdit = fn;
   }
 
   get template() {
     return `<article class="card card--${this._color} ${this._isRepeating() ? `card--repeat` : ``}">
-            <form class="card__form" method="get">
               <div class="card__inner">
                 <div class="card__control">
                   <button type="button" class="card__btn card__btn--edit">edit</button>
@@ -43,8 +34,7 @@ export default class Task {
 
                 <div class="card__color-bar">
                   <svg width="100%" height="10">
-                    <use xlink:href="#wave"></use>
-                  </svg>
+                    <use xlink:href="#wave"></use></svg>
                 </div>
 
                 <div class="card__textarea-wrap">
@@ -55,75 +45,32 @@ export default class Task {
 
                 <div class="card__settings">
                   <div class="card__details">
-                    <div class="card__dates">
-                      <button class="card__date-deadline-toggle" type="button">
-                        date: <span class="card__date-status">${this._dueDate}</span>
-                      </button>
-
-                      <fieldset class="card__date-deadline" disabled>
-                        <label class="card__input-deadline-wrap">
-                          <input class="card__date" type="text" placeholder="23 September" filterName="date" />
-                        </label>
-                        <label class="card__input-deadline-wrap">
-                          <input class="card__time" type="text" placeholder="11:15 PM" filterName="time" />
-                        </label>
-                      </fieldset>
-
-                      <button class="card__repeat-toggle" type="button">
-                        repeat:<span class="card__repeat-status"></span>
-                      </button>
                     <div class="card__hashtag">
                       <div class="card__hashtag-list">${Array.from(this._tags).map((tag) => (`<span class="card__hashtag-inner">
                       <input type="hidden" name="hashtag" value="${tag}" class="card__hashtag-hidden-input" />
                       <button type="button" class="card__hashtag-name">#${tag}</button>
                       <button type="button" class="card__hashtag-delete">delete</button>
-                    </span>`.trim())).join(``)}</div>
-
-                      <label>
-                        <input type="text" class="card__hashtag-input" filterName="hashtag-input" placeholder="Type new hashtag here" />
-                      </label>
+                    </span>`.trim())).join(``)}
                     </div>
-                  </div>
-
-                  <label class="card__img-wrap card__img-wrap--empty">
-                    <input type="file" class="card__img-input visually-hidden" filterName="img" />
-                    <img
-                      src="${this._picture}"
-                      alt="task picture"
-                      class="card__img"
-                    />
-                  </label>
-
-
-                </div>
-
-                <div class="card__status-btns">
-                  <button class="card__save" type="submit">save</button>
-                  <button class="card__delete" type="button">delete</button>
+                    </div>
                 </div>
               </div>
-            </form>
+            </div>
           </article>`.trim();
   }
+
   bind() {
-    this._element.querySelector(`.card__btn--edit`).addEventListener(`click`, this._onEditButtonClick.bind(this));
+    this._element.querySelector(`.card__btn--edit`).addEventListener(`click`, this._onEditButtonClick);
   }
+
   update() {
     if (this._state.isEdit) {
       return this._element.classlistAdd(`card--edit`);
     }
     return this._element.classList.remove(`card--edit`);
   }
-  render() {
-    this._element = createElement(this.template);
-    this.bind();
-    return this._element;
-  }
+
   unbind() {
     this._element.querySelector(`.card__btn--edit`).removeEventListener(`click`, this._onEditButtonClick);
-  }
-  unrender() {
-    this.unbind();
-    this._element = null;
   }
 }
